+++ 3rdParty/libmidi/MIDIProcessorSMF.cpp
@@ -77,12 +77,12 @@ bool MIDIProcessor::ProcessSMF(std::vector<uint8_t> const & data, MIDIContainer 
 
         if (::memcmp(&Data[0], "MTrk", 4) == 0)
         {
-            if (Tail - Data < (ptrdiff_t)8 + ChunkSize)
+            if (Tail - Data < (ptrdiff_t) (8 + ChunkSize))
                 return SetLastErrorCode(MIDIError::InsufficientData);
 
             Data += 8;
 
-            std::vector<uint8_t>::const_iterator ChunkTail = Data + ChunkSize;
+            std::vector<uint8_t>::const_iterator ChunkTail = Data + (int) ChunkSize;
 
             if (!ProcessSMFTrack(Data, ChunkTail, container, true))
                 return false;
@@ -92,7 +92,7 @@ bool MIDIProcessor::ProcessSMF(std::vector<uint8_t> const & data, MIDIContainer 
         // Skip unknown chunks in the stream.
         else
         {
-            if (Tail - Data < (ptrdiff_t)(8) + ChunkSize)
+            if (Tail - Data < (ptrdiff_t) (8 + ChunkSize))
                 return SetLastErrorCode(MIDIError::InsufficientData);
 
             Data += (int64_t)(8) + ChunkSize;
@@ -252,7 +252,7 @@ bool MIDIProcessor::ProcessSMFTrack(std::vector<uint8_t>::const_iterator & data,
             {
                 Buffer.resize((size_t)SysExSize + Size);
 
-                std::copy(data, data + Size, Buffer.begin() + SysExSize);
+                std::copy(data, data + Size, Buffer.begin() + (int) SysExSize);
                 data += Size;
 
                 SysExSize += Size;
