+++ EdMPlayer.h
@@ -24,7 +24,7 @@ public:
 private:
     virtual bool Startup() override;
     virtual void Shutdown() override;
-    virtual void Render(audio_sample *, unsigned long) override;
+    virtual void Render(audio_sample *, uint32_t) override;
 
     virtual void SendEvent(uint32_t) override;
     virtual void SendSysEx(const uint8_t * event, size_t size, uint32_t portNumber) override;
